+++ FE/src/JS/index.js
@@ -1,5 +1,5 @@
 import { initSidebar } from './sidebar/sidebar-app.js';
-import { initTodo } from './todo/app.js';
+import { initTodo } from './todo/todo-app.js';
 
 initTodo();
 initSidebar();
