+++ FE/src/JS/todo/todo-model.js
@@ -4,17 +4,14 @@ export default class Model {
   constructor(workListData) {
     this.userId = workListData.userId;
     this.workList = {};
-    this.categoryId= {};
+    this.categoryId = {};
     this.init(workListData);
   }
 
   init(workListData) {
-    for (let category of workListData.workList) {
+    for (const category of workListData.workList) {
       this.workList[category.categoryName] = category.works;
       this.categoryId[category.categoryName] = category.categoryId;
     }
-    console.log(this.workList)
-    console.log(this.categoryId)
   }
 }
-
